+++ github/trending/api.go
@@ -1,4 +1,4 @@
-package github
+package trending
 
 import (
 	"fmt"
