+++ github/trending/api_test.go
@@ -1,4 +1,4 @@
-package github
+package trending
 
 import (
 	"testing"
